+++ CHANGES.rst
@@ -6,7 +6,7 @@ Changelog of lizard-ui
 -----------------
 
 - Fix layout bug in login modal. The margin was too large.
-
+- Enhance  404 and 500 html error templates.
 
 4.28 (2013-05-14)
 -----------------